fix: search and redundant wrappers
