+++ client/package.json
@@ -34,10 +34,10 @@
   },
   "devDependencies": {
     "@svgr/webpack": "4.1.0",
-    "@types/jest": "25.1.2",
+    "@types/jest": "25.1.3",
     "@types/lodash": "4.14.149",
-    "@types/node": "13.7.1",
-    "@types/react": "16.9.19",
+    "@types/node": "13.7.4",
+    "@types/react": "16.9.22",
     "@types/react-dom": "16.9.5",
     "babel-eslint": "9.0.0",
     "babel-jest": "23.6.0",
@@ -65,7 +65,7 @@
     "redux-mock-store": "1.5.4",
     "sortpack": "2.1.1",
     "terser-webpack-plugin": "1.2.2",
-    "typescript": "3.7.5",
+    "typescript": "3.8.2",
     "webpack": "4.41.6",
     "webpack-dev-server": "3.10.3",
     "webpack-manifest-plugin": "2.2.0",
