+++ client/src/components/resource/tags.tsx
@@ -1,30 +1,19 @@
 import React from 'react'
 import styled from 'styled-components/macro'
 
-const Wrapper = styled.div`
+const Span = styled.span`
   font-size: 12px;
   margin-top: auto;
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
-
-  & > * {
-    margin-right: 4px;
-  }
-
-  & > span {
-    color: ${(props) => props.theme.mutedText};
-  }
+  color: ${(props) => props.theme.mutedText};
 `
 
 type Props = {
   tags: string[]
 }
 
-const Tags = (props: Props) => (
-  <Wrapper>
-    <span>Tags: {props.tags.join(' ')}</span>
-  </Wrapper>
-)
+const Tags = (props: Props) => <Span>Tags: {props.tags.join(' ')}</Span>
 
 export default Tags
