+++ server/package.json
@@ -23,19 +23,19 @@
     "js-yaml": "3.13.1",
     "koa": "2.11.0",
     "koa-mid": "3.0.5",
-    "koa-simple-static": "4.0.2",
+    "koa-simple-static": "4.0.3",
     "lodash": "4.17.15",
     "lowdb": "1.0.0",
     "pluralize": "8.0.0"
   },
   "devDependencies": {
-    "@types/jest": "25.1.2",
+    "@types/jest": "25.1.3",
     "@types/js-yaml": "3.12.2",
     "@types/koa": "2.11.1",
     "@types/koa-bodyparser": "4.3.0",
     "@types/koa-compress": "2.0.9",
     "@types/koa__router": "8.0.2",
-    "@types/node": "13.7.1",
+    "@types/node": "13.7.4",
     "@types/pluralize": "0.0.29",
     "@types/supertest": "2.0.8",
     "coveralls": "3.0.9",
@@ -49,9 +49,9 @@
     "prettier": "1.19.1",
     "sortpack": "2.1.1",
     "supertest": "4.0.2",
-    "ts-jest": "25.2.0",
+    "ts-jest": "25.2.1",
     "ts-node": "8.6.2",
-    "typescript": "3.7.5"
+    "typescript": "3.8.2"
   },
   "jest": {
     "collectCoverageFrom": [
